refactor(InputView, Application): move input collection and validation logic to InputView
Move all input-related logic from the main function in Application.kt to the InputView class:
- this includes methods for reading and validating the purchase amount, winning numbers, and bonus number
